+++ cmd/cronosd/cmd/root.go
@@ -150,6 +150,7 @@ func initRootCmd(rootCmd *cobra.Command, encodingConfig params.EncodingConfig) {
 
 	// add rosetta
 	rootCmd.AddCommand(server.RosettaCommand(encodingConfig.InterfaceRegistry, encodingConfig.Codec))
+	rootCmd.AddCommand(RestoreSnapshotCmd(a.newApp, app.DefaultNodeHome))
 }
 
 func addModuleInitFlags(startCmd *cobra.Command) {